Converted plugin message packet
